+++ src/voice-orchestrator-agent/VoiceLiveAgentExtensions.cs
@@ -1,3 +1,4 @@
+using System.Text.RegularExpressions;
 using Azure.AI.VoiceLive;
 using Microsoft.Agents.AI;
 
@@ -8,7 +9,7 @@ namespace VoiceOrchestratorAgent;
 /// Analogous to <c>agent.AsAIFunction()</c> in Microsoft Agent Framework, but targeting
 /// the Voice Live SDK's <see cref="VoiceLiveFunctionDefinition"/> format.
 /// </summary>
-public static class VoiceLiveAgentExtensions
+public static partial class VoiceLiveAgentExtensions
 {
     private static readonly BinaryData s_queryParameters = BinaryData.FromObjectAsJson(new
     {
@@ -31,10 +32,20 @@ public static class VoiceLiveAgentExtensions
     /// </summary>
     public static VoiceLiveFunctionDefinition AsVoiceLiveTool(this AIAgent agent)
     {
-        return new VoiceLiveFunctionDefinition(agent.Name)
+        return new VoiceLiveFunctionDefinition(SanitizeAgentName(agent.Name))
         {
             Description = agent.Description ?? $"Invoke the {agent.Name} agent",
             Parameters = s_queryParameters
         };
     }
+
+    private static string? SanitizeAgentName(string? agentName)
+    {
+        return agentName is null
+            ? agentName
+            : InvalidNameCharsRegex().Replace(agentName, "_");
+    }
+
+    [GeneratedRegex("[^0-9A-Za-z]+")]
+    private static partial Regex InvalidNameCharsRegex();
 }
